Docs: upgrade Docusaurus from 2.4.3 to 3.7.0 with TypeScript support
- Convert all JavaScript files to TypeScript (.js to .tsx/.ts), add proper type annotations, and migrate homepage and version pages to use TypeScript with type-safe imports and components
- Automatically generate `latestVersion.json` containing the latest version info, and it's used to get the current version.
