+++ website/.gitignore
@@ -20,3 +20,4 @@ yarn-debug.log*
 yarn-error.log*
 
 /versioned_docs
+/latestVersion.json
