+++ website/src/components/HomepageFeatures/styles.module.css
@@ -1,0 +1,16 @@
+.features {
+  display: flex;
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+}
+
+.featureSvg {
+  height: 200px;
+  width: 200px;
+}
+
+.featurePng {
+  height: 200px;
+  width: 200px;
+}
