+++ website/src/pages/versionsArchived.json
@@ -1,5 +1,9 @@
 [
   {
+    "name": "2.0.0",
+    "label": "2.0.0"
+  },
+  {
     "name": "2.0.0-beta14",
     "label": "2.0.0-beta14"
   },
