+++ website/src/types/commonTypes.ts
@@ -1,0 +1,11 @@
+
+
+interface ArchivedVersion {
+  name: string;
+  label: string;
+  path?: string;
+}
+
+interface LatestVersion {
+  version: string;
+}
