+++ website/versioned_sidebars/version-v1-sidebars.json
@@ -1,8 +1,8 @@
 {
-  "version-v1/docs": [
+  "sidebar": [
     {
       "type": "autogenerated",
-      "dirName": "."
+      "dirName": "docs"
     }
   ]
 }
